moved get_default_post_processing_commands()
