+++ controllers/examples/magnebot/minimal_scene_setup.py
@@ -2,7 +2,8 @@ from tdw.controller import Controller
 from tdw.tdw_utils import TDWUtils
 from tdw.add_ons.step_physics import StepPhysics
 from tdw.add_ons.object_manager import ObjectManager
-from magnebot import Magnebot, MagnebotController
+from magnebot import Magnebot
+from magnebot.util import get_default_post_processing_commands
 
 
 """
@@ -21,7 +22,7 @@ commands = [{"$type": "load_scene",
 commands.extend(c.get_add_physics_object(model_name="rh10",
                                          position={"x": -2, "y": 0, "z": -1.5},
                                          object_id=c.get_unique_id()))
-commands.extend(MagnebotController.get_default_post_processing_commands())
+commands.extend(get_default_post_processing_commands())
 c.communicate(commands)
 print(magnebot.dynamic.transform.position)
 for object_id in objects.transforms:
