+++ controllers/examples/magnebot_controller/collision_detection.py
@@ -1,5 +1,6 @@
 from tdw.tdw_utils import TDWUtils
 from magnebot import MagnebotController
+from magnebot.util import get_default_post_processing_commands
 
 
 class CollisionDetection(MagnebotController):
@@ -17,7 +18,8 @@ class CollisionDetection(MagnebotController):
         self._init_scene(scene=scene,
                          objects=objects,
                          position={"x": 1, "y": 0, "z": -3},
-                         rotation={"x": 0, "y": 46, "z": 0})
+                         rotation={"x": 0, "y": 46, "z": 0},
+                         post_processing=get_default_post_processing_commands())
 
     def run(self, arrived_offset: float, objects: bool) -> None:
         self.init_scene()
