+++ doc/manual/magnebot/scene_setup.md
@@ -48,12 +48,13 @@ To replicate `MagnebotController.init_scene()` we need to add a few more things:
 
 This controller adds a Magnebot and an object to a scene:
 
-```## Add Magnebot to a floorplan scenepython
+```python
 from tdw.controller import Controller
 from tdw.tdw_utils import TDWUtils
 from tdw.add_ons.step_physics import StepPhysics
 from tdw.add_ons.object_manager import ObjectManager
-from magnebot import Magnebot, MagnebotController
+from magnebot import Magnebot
+from magnebot.util import get_default_post_processing_commands
 
 c = Controller()
 magnebot = Magnebot()
@@ -67,7 +68,7 @@ commands = [{"$type": "load_scene",
 commands.extend(c.get_add_physics_object(model_name="rh10",
                                          position={"x": -2, "y": 0, "z": -1.5},
                                          object_id=c.get_unique_id()))
-commands.extend(MagnebotController.get_default_post_processing_commands())
+commands.extend(get_default_post_processing_commands())
 c.communicate(commands)
 print(magnebot.dynamic.transform.position)
 for object_id in objects.transforms:
@@ -143,7 +144,6 @@ class ResetScene(Controller):
         self.communicate([{"$type": "load_scene",
                            "scene_name": "ProcGenScene"},
                           TDWUtils.create_empty_room(12, 12)])
-
 ```
 
 ***
